# backend/openapi_full.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Voice Assistant MCP Server",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/agent/orchestrate": {
      "post": {
        "summary": "Orchestrate Agent",
        "description": "Process a query using the agent orchestrator with function calling.",
        "operationId": "orchestrate_agent_api_agent_orchestrate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QueryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/agent/stream": {
      "post": {
        "tags": [
          "agent"
        ],
        "summary": "Stream Query",
        "description": "Stream a response from the agent.\n\nReturns a streaming response that yields text chunks as they're generated.\nNote: Tool calling happens before streaming begins.",
        "operationId": "stream_query_api_agent_stream_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentQuery"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/agent/tools/chart/search": {
      "get": {
        "tags": [
          "agent"
        ],
        "summary": "Search Chart Tools",
        "description": "Search for chart tools using semantic search.\n\nArgs:\n    query: Natural language query (e.g., \"show momentum indicators\")\n    top_k: Maximum number of results to return\n\nReturns:\n    List of relevant chart tools with knowledge base context",
        "operationId": "search_chart_tools_api_agent_tools_chart_search_get",
        "parameters": [
          {
            "name": "query",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Query"
            }
          },
          {
            "name": "top_k",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 5,
              "title": "Top K"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/agent/test/toggle-education-mode": {
      "post": {
        "tags": [
          "agent"
        ],
        "summary": "Toggle Education Mode",
        "description": "Toggle education mode between LLM and templates (for A/B testing).",
        "operationId": "toggle_education_mode_api_agent_test_toggle_education_mode_post",
        "parameters": [
          {
            "name": "enabled",
            "in": "query",
            "required": true,
            "schema": {
              "type": "boolean",
              "title": "Enabled"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/agent/analyze-chart": {
      "post": {
        "tags": [
          "agent"
        ],
        "summary": "Analyze Chart",
        "description": "Analyze a chart image using the vision model-based analyzer.",
        "operationId": "analyze_chart_api_agent_analyze_chart_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChartImageRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/agent/chart-snapshot": {
      "post": {
        "tags": [
          "agent"
        ],
        "summary": "Ingest Chart Snapshot",
        "description": "Ingest a chart snapshot and optionally trigger automated analysis.",
        "operationId": "ingest_chart_snapshot_api_agent_chart_snapshot_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChartSnapshotIngestRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChartSnapshotResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/agent/chart-snapshot/{symbol}": {
      "get": {
        "tags": [
          "agent"
        ],
        "summary": "Get Latest Chart Snapshot",
        "description": "Return the latest chart snapshot for a symbol/timeframe.",
        "operationId": "get_latest_chart_snapshot_api_agent_chart_snapshot__symbol__get",
        "parameters": [
          {
            "name": "symbol",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Symbol"
            }
          },
          {
            "name": "timeframe",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Timeframe"
            }
          },
          {
            "name": "include_image",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Include Image"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/ChartSnapshotResponse"
                    },
                    {
                      "type": "null"
                    }
                  ],
                  "title": "Response Get Latest Chart Snapshot Api Agent Chart Snapshot  Symbol  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/agent/voice-query": {
      "post": {
        "tags": [
          "agent"
        ],
        "summary": "Process Voice Query",
        "description": "Process a voice query: agent processes it and sends response to TTS.\n\nThis is the integration point between the agent and voice interface:\n1. Receives transcript from STT\n2. Agent processes query with tools\n3. Sends response text to TTS via Realtime API\n\nReturns the agent response data (text version).",
        "operationId": "process_voice_query_api_agent_voice_query_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentQuery"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/agent/pattern-verdict": {
      "post": {
        "tags": [
          "agent"
        ],
        "summary": "Submit Pattern Verdict",
        "description": "Submit analyst verdict for a pattern.\n\nThis endpoint:\n1. Records the analyst's decision (accept/reject/defer)\n2. Updates pattern state in the system\n3. Triggers pattern overlay updates via WebSocket\n4. Creates audit trail for compliance",
        "operationId": "submit_pattern_verdict_api_agent_pattern_verdict_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatternVerdict"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PatternVerdictResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AgentQuery": {
        "properties": {
          "query": {
            "type": "string",
            "title": "Query"
          },
          "conversation_history": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Conversation History"
          },
          "stream": {
            "type": "boolean",
            "title": "Stream",
            "default": false
          },
          "session_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Id"
          },
          "user_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Id"
          }
        },
        "type": "object",
        "required": [
          "query"
        ],
        "title": "AgentQuery",
        "description": "Request model for agent queries."
      },
...
      "ChartImageRequest": {
        "properties": {
          "image_base64": {
            "type": "string",
            "title": "Image Base64"
          },
          "context": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "image_base64"
        ],
        "title": "ChartImageRequest",
        "description": "Request model for chart image analysis."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "QueryRequest": {
        "properties": {
          "query": {
            "type": "string",
            "title": "Query"
          },
          "session_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Id"
          },
          "user_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Id"
          },
          "include_history": {
            "type": "boolean",
            "title": "Include History",
            "default": true
          },
          "voice_enabled": {
            "type": "boolean",
            "title": "Voice Enabled",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "query"
        ],
        "title": "QueryRequest",
        "description": "Request model for voice queries."
      },
...
    }
  }
}